Change Simperium data model to user based system w/ tie-ins on Liveblog status change
Data was originally stored in Simperium using a single username/password
and a data bucket per-post. Since there are limits on the number of
buckets per user, the plugin will now create a new username per post and
store data under the "feed" bucket. Additionally, the plugin will now
try to pre-populate Simperium when Liveblog is enabled for a post, and
will delete the Simperium user when Liveblog is archived or the post is
deleted.