Fix IE XHR latest timestamp bug
In the event that the same XHR request gets sent twice (mainly on
scroll), IE tends to just reply with a 304 status (not modified) and
doesn't supply a new response date which throws off the entire
calculation system. In such cases we now default to the previous known
server time.