Change remove_replaced_entries to not verify replaced exists
remove_replaced_entries only removed an entry if the replacement item
existed in the collection. If an entry was updated prior to being
deleted then the "update" lingered in the database and the entry would
keep showing up on page refresh because the original (the replaced
entry) had been deleted. Removing this logic check ensures that only
original entries show up in the feed.