Add mode parameter to remove_replace_entries
remove_replaced_entries only removed an entry if the replacement item
existed in the collection which caused problems on page load when all
entries are handled. Since deleting an entry only removes the original,
any update records are left lingering and would display since the
original could not be found. Added a "mode" parameter to help
differentiate between the two states.